Сделал сервис для  Airport(только вывод списком и дефолтные данные). Осталось узнать на счёт модели от человека и остальных методах. По теор. они не нужны. Иницилизация данных в 1 файле. Глянуть структуру папок (тк маршрутизация столько моими папками громоздка)

+++ UtopiaCity/Common/Initializers/TimelineEventInitializer.cs
@@ -14,7 +14,15 @@ namespace UtopiaCity.Common.Initializers
                 return;
             }
 
+            if (!context.PermitedModel.Any())
+            {
+                return;
+            }
+
             context.RemoveRange(context.TimelineModel.ToList());
+            
+            context.RemoveRange(context.PermitedModel.ToList());
+
             context.SaveChanges();
         }
 
@@ -71,6 +79,32 @@ namespace UtopiaCity.Common.Initializers
                     }
                     );
             }
+
+            //Permited Conditons
+            if(!context.PermitedModel.Any())
+            {
+                context.PermitedModel.AddRange(
+                    new Models.TimelineModel.PermitedModel
+                    {
+                        Season = "Winter",
+                        SpeedOfWind = 10.01M,
+                        GovernmentStatus = false,
+                        Rainfall = "rain",
+                        TimeOfDay = "night"
+                    },
+
+                    new Models.TimelineModel.PermitedModel
+                    {
+                        Season = "Winter",
+                        SpeedOfWind = 10.01M,
+                        GovernmentStatus = false,
+                        Rainfall = "rain",
+                        TimeOfDay = "night"
+                    }
+
+                    );
+            }
+
 
             context.SaveChanges();
         }
